+++ src/extension-opera/skipto.js
@@ -227,7 +227,7 @@
   const cssMenuTemplate = document.createElement('template');
   cssMenuTemplate.textContent = `
 $skipToId.popup {
-  top: -30px;
+  top: -36px;
   transition: top 0.35s ease;
 }
 
@@ -302,10 +302,6 @@ $skipToId button {
     transition: top 0.35s ease;
     display: none;
   }
-
-  $skipToId.popup {
-    top: -26px;
-  }
 }
 
 @media screen and (min-width: $smallBreakPointpx) and (max-width: $mediumBreakPointpx) {
@@ -330,11 +326,6 @@ $skipToId button {
     transition: top 0.35s ease;
     display: none;
   }
-
-  $skipToId.popup {
-    top: -28px;
-  }
-
 }
 
 $skipToId.static {